requirements.txt
flask
flask-jwt-extended
python-dotenv
requests
sqlalchemy
bcrypt
openai
gradio_client
